requirements.txt
flask
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pyodbc
gunicorn